Use MultiConformanceChecker to get conformances to be "complete"
